added engraving score reader
