+++ src/engraving/rw/compat/readchordlisthook.cpp
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-#include "chordlist.h"
+#include "readchordlisthook.h"
 
 #include <QDebug>
